refactor test suites, implement more tests
