+++ tests/js/NedbClientCREATE.test.js
@@ -1,0 +1,60 @@
+const {SuperCamo, CollectionListEntry, NedbClient, CollectionAccessError} = require("../../dist/index.js");
+// const {} = require("../../dist/structures/NedbClient.js");
+
+const {describe, test, expect} = require("@jest/globals");
+const { User } = require("./helpers/TestDocumentDefs.js");
+
+const firstUserData = {
+	email: "test@email.com",
+	bio: {tagline: "Test user.", blurb: "Super cool test user with a long, descriptive blurb."}
+};
+
+const secondUserData = {
+	email: "test2@email.com",
+	bio: {tagline: "Test user, the second.", blurb: "Another super cool test user with a long, descriptive blurb."}
+};
+
+/**
+ * @type {NedbClient}
+ */
+let newClient;
+
+afterAll(async () => {
+	await newClient.dropDatabase();
+	SuperCamo.clientDelete("NedbClientClassCreateTestDb");
+});
+
+
+beforeAll(async () => {
+	newClient = SuperCamo.clientConnect(
+		"NedbClientClassCreateTestDb",
+		"./.testing/NedbClientClassCreateTestDb/",
+		[
+			new CollectionListEntry("Users", User)
+		]
+	);
+
+	await newClient.insertOne("Users", firstUserData);
+
+	await newClient.insertOne("Users", secondUserData);
+})
+
+describe.skip("Database can perform CREATE operations", () => {
+	describe("insertOne (object) operations", () => {
+		test("returns an appropriate object when given valid data for database insertion", async () => {
+
+		});
+	});
+
+	describe("insertMany (objects) operations", () => {
+
+	});
+
+	describe("createOne (document) operations", () => {
+
+	});
+
+	describe("createMany (documents) operations", () => {
+
+	});
+});
