+++ tests/js/NedbClientUPDATE.test.js
@@ -1,0 +1,58 @@
+const {SuperCamo, CollectionListEntry, NedbClient, CollectionAccessError} = require("../../dist/index.js");
+// const {} = require("../../dist/structures/NedbClient.js");
+
+const {describe, test, expect} = require("@jest/globals");
+const { User } = require("./helpers/TestDocumentDefs.js");
+
+const firstUserData = {
+	email: "test@email.com",
+	bio: {tagline: "Test user.", blurb: "Super cool test user with a long, descriptive blurb."}
+};
+
+const secondUserData = {
+	email: "test2@email.com",
+	bio: {tagline: "Test user, the second.", blurb: "Another super cool test user with a long, descriptive blurb."}
+};
+
+/**
+ * @type {NedbClient}
+ */
+let newClient;
+
+afterAll(async () => {
+	await newClient.dropDatabase();
+	SuperCamo.clientDelete("NedbClientClassUpdateTestDb");
+});
+
+
+beforeAll(async () => {
+	newClient = SuperCamo.clientConnect(
+		"NedbClientClassUpdateTestDb",
+		"./.testing/NedbClientClassUpdateTestDb/",
+		[
+			new CollectionListEntry("Users", User)
+		]
+	);
+
+	await newClient.insertOne("Users", firstUserData);
+
+	await newClient.insertOne("Users", secondUserData);
+})
+
+describe.skip("Database can perform UPDATE operations", () => {
+	describe("findAndUpdateOneDocument operations", () => {
+		test("returns null when no data matches the given query", async () => {
+			
+		})
+	});
+	describe("findAndUpdateOneObject operations", () => {
+
+	});
+
+	describe("findAndUpdateManyDocuments operations", () => {
+
+	});
+	describe("findAndUpdateManyObjects operations", () => {
+
+	});
+});
